add FIleStorage

+++ teddy_blue_api/src/main/java/com/teddy_blue/Services/AnnonceServiceImpl.java
@@ -2,12 +2,14 @@ package com.teddy_blue.Services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
-import javax.transaction.Transactional;
+import javax.persistence.EntityNotFoundException;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.teddy_blue.Entities.Annonce;
 import com.teddy_blue.Entities.Category;
@@ -22,22 +24,21 @@ import com.teddy_blue.dtos.AnnonceItem;
 import com.teddy_blue.dtos.AnnonceUpdate;
 
 @Service
+@Transactional
 public class AnnonceServiceImpl implements AnnonceService {
 
     private final AnnonceRepository annonceRepository;
     private final UserRepository userRepository;
     private final CategoryRepository categoryRepository;
     private final SharingMethodRepository sharingMethodRepository;
+    private final FileStorage storage;
 
-    @Autowired
-    public AnnonceServiceImpl(AnnonceRepository annonceRepository,
-                              UserRepository userRepository,
-                              CategoryRepository categoryRepository,
-                              SharingMethodRepository sharingMethodRepository) {
+    public AnnonceServiceImpl(FileStorage storage, AnnonceRepository annonceRepository, UserRepository userRepository, CategoryRepository categoryRepository, SharingMethodRepository sharingMethodRepository) {
         this.annonceRepository = annonceRepository;
         this.userRepository = userRepository;
         this.categoryRepository = categoryRepository;
         this.sharingMethodRepository = sharingMethodRepository;
+        this.storage = storage;
     }
     
     private AnnonceItem convertToAnnonceItem(Annonce annonce) {
@@ -62,22 +63,21 @@ public class AnnonceServiceImpl implements AnnonceService {
         return item;
     }
 
-    @Transactional // read only = false
     @Override
-    public void createAnnonce (AnnonceCreate inputs) {
+    public void createAnnonce(AnnonceCreate inputs) {
         Annonce entity = new Annonce();
         entity.setTitle(inputs.getTitle());
         entity.setTextAnnonce(inputs.getTextAnnonce());
-        entity.setPhotoLink(inputs.getPhotoLink());
-        User user = userRepository.findById(inputs.getUserId())
-                         .orElseThrow(() -> new RuntimeException("User not found"));
-        entity.setUser(user);
-        SharingMethod sharingMethod = sharingMethodRepository.findById(inputs.getSharingMethodId())
-                                   .orElseThrow(() -> new RuntimeException("SharingMethod not found"));
-        entity.setSharingMethod(sharingMethod);
-        Category category = categoryRepository.findById(inputs.getCategoryId())
-                            .orElseThrow(() -> new RuntimeException("Category not found"));
-        entity.setCategory(category);
+        MultipartFile file = inputs.getFile();
+	String baseName = UUID.randomUUID().toString();
+	String fileName = storage.store(file, baseName);
+	entity.setPhotoLink(fileName);
+        Optional<User> userOpt = userRepository.findById(inputs.getUserId());
+        userOpt.ifPresent(entity::setUser);
+        Optional<SharingMethod> sharingMethodOpt = sharingMethodRepository.findById(inputs.getSharingMethodId());
+        sharingMethodOpt.ifPresent(entity::setSharingMethod);
+        Optional<Category> categoryOpt = categoryRepository.findById(inputs.getCategoryId());
+        categoryOpt.ifPresent(entity::setCategory);
         annonceRepository.save(entity);
     }
 
@@ -97,33 +97,45 @@ public class AnnonceServiceImpl implements AnnonceService {
     @Transactional
     @Override
     public void updateAnnonce(Long id, AnnonceUpdate inputs) {
-        Annonce entity = annonceRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Annonce not found"));
-
-        entity.setTitle(inputs.getTitle());
-        entity.setTextAnnonce(inputs.getTextAnnonce());
-        entity.setPhotoLink(inputs.getPhotoLink());
-
-        if (inputs.getCategoryId() != null) {
-            Category category = categoryRepository.findById(inputs.getCategoryId())
-                .orElseThrow(() -> new RuntimeException("Category not found"));
-            entity.setCategory(category);
+        Optional<Annonce> annonceOpt = annonceRepository.findById(id);
+        if (annonceOpt.isPresent()) {
+            Annonce entity = annonceOpt.get();
+            entity.setTitle(inputs.getTitle());
+            entity.setTextAnnonce(inputs.getTextAnnonce());
+            MultipartFile file = inputs.getFile();
+            if (file != null) {
+                String original = entity.getPhotoLink();
+                String baseName = UUID.randomUUID().toString();
+                String newFullName = storage.replace(file, baseName, original);
+                entity.setPhotoLink(newFullName);
+            }
+            Optional<User> userOpt = userRepository.findById(inputs.getUserId());
+            userOpt.ifPresent(entity::setUser);
+            Optional<SharingMethod> sharingMethodOpt = sharingMethodRepository.findById(inputs.getSharingMethodId());
+            sharingMethodOpt.ifPresent(entity::setSharingMethod);
+            Optional<Category> categoryOpt = categoryRepository.findById(inputs.getCategoryId());
+            categoryOpt.ifPresent(entity::setCategory);
+            annonceRepository.save(entity);
         }
-
-        if (inputs.getSharingMethodId() != null) {
-            SharingMethod sharingMethod = sharingMethodRepository.findById(inputs.getSharingMethodId())
-                .orElseThrow(() -> new RuntimeException("SharingMethod not found"));
-            entity.setSharingMethod(sharingMethod);
+        else {
+            throw new EntityNotFoundException("Annonce with ID " + id + " not found");
         }
-
-        annonceRepository.save(entity);
     }
 
     @Transactional
     @Override
     public void deleteAnnonce(Long id) {
-        Annonce entity = annonceRepository.findById(id).orElseThrow(() -> new RuntimeException("Annonce not found"));
-        annonceRepository.delete(entity);
+        Optional<Annonce> annonceOpt = annonceRepository.findById(id);
+        if (annonceOpt.isPresent()) {
+            Annonce entity = annonceOpt.get();
+            String fullName = entity.getPhotoLink();
+            annonceRepository.delete(entity);
+            storage.delete(fullName);
+        }
+        // Handle case where Annonce is not found
+        else {
+            throw new EntityNotFoundException("Annonce with ID " + id + " not found");
+        }
     }
-
 }
+
